+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdkVersion 26
     buildToolsVersion '27.0.1'
     defaultConfig {
-        applicationId "com.olq.whatsthatflag"
+        applicationId "me.wozappz.whatsthatflag"
         minSdkVersion 16
         targetSdkVersion 26
         versionCode 1
