--- app/src/main/java/com/olq/whatsthatflag/App.kt
@@ -1,20 +1,0 @@
-package com.olq.whatsthatflag
-
-import android.app.Application
-import com.squareup.leakcanary.LeakCanary
-
-/**
- * Created by olq on 10.12.17.
- */
-class App : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-
-        if (LeakCanary.isInAnalyzerProcess(this)) {
-            return
-        }
-
-        LeakCanary.install(this)
-    }
-}
