--- app/src/main/java/com/olq/whatsthatflag/screens/BaseView.kt
@@ -1,9 +1,0 @@
-package com.olq.whatsthatflag.screens
-
-/**
- * Created by olq on 20.11.17.
- */
-interface BaseView<T> {
-
-    var presenter: T
-}
