--- app/src/main/java/com/olq/whatsthatflag/screens/menu/ContinentEnum.kt
@@ -1,13 +1,0 @@
-package com.olq.whatsthatflag.screens.menu
-
-/**
- * Created by olq on 10.12.17.
- */
-enum class CONTINENT {
-    GLOBAL,
-    EUROPE,
-    ASIA,
-    AMERICAS,
-    AFRICA,
-    OCEANIA
-}
