+++ app/src/main/java/me/wozappz/whatsthatflag/screens/BaseView.kt
@@ -1,0 +1,9 @@
+package me.wozappz.whatsthatflag.screens
+
+/**
+ * Created by olq on 20.11.17.
+ */
+interface BaseView<T> {
+
+    var presenter: T
+}
