+++ app/src/main/java/me/wozappz/whatsthatflag/screens/menu/ContinentEnum.kt
@@ -1,0 +1,13 @@
+package me.wozappz.whatsthatflag.screens.menu
+
+/**
+ * Created by olq on 10.12.17.
+ */
+enum class CONTINENT {
+    GLOBAL,
+    EUROPE,
+    ASIA,
+    AMERICAS,
+    AFRICA,
+    OCEANIA
+}
