+++ app/src/main/res/layout-sw720dp/activity_game.xml
@@ -7,7 +7,7 @@
     android:layout_gravity="center"
     android:maxHeight="1080dp"
     android:maxWidth="720dp"
-    tools:context="com.olq.whatsthatflag.screens.game.GameActivity"
+    tools:context="me.wozappz.whatsthatflag.screens.game.GameActivity"
     tools:layout_editor_absoluteX="0dp"
     tools:layout_editor_absoluteY="73dp">
