+++ app/src/main/res/layout/activity_start.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.olq.whatsthatflag.screens.start.StartActivity"
+    tools:context="me.wozappz.whatsthatflag.screens.start.StartActivity"
     tools:layout_editor_absoluteX="0dp"
     tools:layout_editor_absoluteY="0dp">
