+++ app/src/main/res/layout/radio_group_table_layout.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.olq.whatsthatflag.screens.menu.RadioGroupTableLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<me.wozappz.whatsthatflag.screens.menu.RadioGroupTableLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content">
 
@@ -73,4 +73,4 @@
             android:theme="@style/CustomRadioButtonTheme" />
     </TableRow>
 
-</com.olq.whatsthatflag.screens.menu.RadioGroupTableLayout>
+</me.wozappz.whatsthatflag.screens.menu.RadioGroupTableLayout>
